# Extract system test setup to system_helper