Defining these classes in abstract_unit makes running all Action Pack tests slower because they have to download/setup browsers for system tests (even if the tests being run never end up using the system test classes).

This commit moves the system test class definitions into a helper which is only included in the two test files that use them. This will make running non-system Action Pack test files much much faster.