@yahonda 2025-04-01
> This will make running non-system Action Pack test files much much faster.

Hi, Do you have any benchmarks?