@skipkayhil 2025-04-01
On main:

```
❯ time bin/test test/dispatch/cookies_test.rb
bin/test test/dispatch/cookies_test.rb  1.37s user 0.49s system 108% cpu 1.719 total
```

This branch
```
❯ time bin/test test/dispatch/cookies_test.rb
bin/test test/dispatch/cookies_test.rb  1.20s user 0.37s system 119% cpu 1.317 total
```

Not super scientific but these times were pretty consistent for me when switching between branches. The real speedup is when a driver is downloaded: my first test run took 14 seconds to download a driver even though `cookies_test` doesn't need one.