@yahonda 2025-04-01
That makes sense. 

> The real speedup is when a driver is downloaded: my first test run took 14 seconds to download a driver even though cookies_test doesn't need one.